% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws.secretsmanager-package.R
\docType{package}
\name{aws.secretsmanager-package}
\alias{aws.secretsmanager-package}
\alias{aws.secretsmanager}
\title{aws.secretsmanager}
\description{
AWS Secrets Manager Client Package
}
\details{
A simple client package for the Amazon Web Services (AWS) Secrets Manager REST API.
}
\author{
Jack Vanlightly <jack.vanlightly@gmail.com>
}
\keyword{package}
