% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cancel_rotate_secret.R
\name{cancel_rotate_secret}
\alias{cancel_rotate_secret}
\title{Cancel Rotate Secret}
\usage{
cancel_rotate_secret(secret_id, verbose = getOption("verbose", FALSE))
}
\arguments{
\item{secret_id}{The character string ID of the secret to retrieve}

\item{verbose}{TRUE to print out debug information. Defaults to FALSE}
}
\value{
The meta data about the secret stored in Secrets Manager (Name, ARN, Version)
}
\description{
AWS Secrets Manager - Disables automatic scheduled rotation and cancels the rotation of a secret if one is currently in progress.
}
\details{
Disables automatic scheduled rotation and cancels the rotation of a secret if one is currently in progress.
}
