% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_secret_string.R
\name{create_secret_string}
\alias{create_secret_string}
\title{Create Secret String}
\usage{
create_secret_string(secret_id, secret_string,
  description = getOption("description", ""), cmk = getOption("cmk_id",
  ""), tags = getOption("tags", data.frame()),
  verbose = getOption("verbose", FALSE))
}
\arguments{
\item{secret_id}{The character string ID of the secret to create. It should be the friendly name of the new secret. Minimum length of 1. Maximum length of 512.}

\item{secret_string}{The character string secret text to be stored and encrypted. To encrypt multiple values, it is recommended to use use a JSON text string argument and specify key/value pairs.}

\item{description}{(Optional) The character string description of the secret. Maximum length of 2048}

\item{cmk}{(Optional) Character string that specifies the ARN, Key ID, or alias of the AWS KMS customer master key (CMK) to be used to encrypt the SecretString or SecretBinary values in the versions stored in this secret.}

\item{tags}{(Optional) A data frame with Key and Value fields that specifies a list of user-defined tags that are attached to the secret. Each tag is a "Key" and "Value" pair of strings. This operation only appends tags to the existing list of tags.}

\item{verbose}{TRUE to print out debug information. Defaults to FALSE}
}
\value{
The meta data about the secret stored in Secrets Manager
}
\description{
AWS Secrets Manager Create Secret String
}
\details{
Creates a new secret. A secret in Secrets Manager consists of both the protected secret data and the important information needed to manage the secret.
}
