% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_resource_policy.R
\name{delete_resource_policy}
\alias{delete_resource_policy}
\title{Delete Resource Policy}
\usage{
delete_resource_policy(secret_id, verbose = getOption("verbose", FALSE))
}
\arguments{
\item{secret_id}{The Secret ID of the secret that you want to delete the attached resource-based policy for.}

\item{verbose}{TRUE to print out debug information. Defaults to FALSE}
}
\value{
The meta data about the secret stored in Secrets Manager
}
\description{
AWS Secrets Manager Delete Resource Policy
}
\details{
Deletes the resource-based permission policy that's attached to the secret.
}
