% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_secret.R
\name{delete_secret}
\alias{delete_secret}
\title{Delete Secret}
\usage{
delete_secret(secret_id, force_delete_without_recovery, recovery_window,
  verbose = getOption("verbose", FALSE))
}
\arguments{
\item{secret_id}{The character string ID of the secret to retrieve}

\item{force_delete_without_recovery}{(Optional) A boolean that specifies that the secret is to be deleted without any recovery window. Defaults to FALSE}

\item{recovery_window}{(Optional) An integer that specifies the number of days that Secrets Manager waits before it can delete the secret. You can't use both this parameter and the ForceDeleteWithoutRecovery parameter in the same API call. This value can range from 7 to 30 days. The default value is 30.}

\item{verbose}{(Optional) TRUE to print out debug information. Defaults to FALSE}
}
\value{
The meta data about the secret stored in Secrets Manager
}
\description{
AWS Secrets Manager Delete Secret
}
\details{
Deletes an entire secret and all of its versions. You can optionally include a recovery window during which you can restore the secret. If you don't specify a recovery window value, the operation defaults to 30 days.
}
