% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_secret.R
\name{describe_secret}
\alias{describe_secret}
\title{Describe Secret}
\usage{
describe_secret(secret_id, verbose = getOption("verbose", FALSE))
}
\arguments{
\item{string}{The ID of the secret to retrieve}

\item{boolean}{TRUE to print out debug information. Defaults to FALSE}
}
\value{
The meta data about the secret stored in Secrets Manager
}
\description{
AWS Secrets Manager Describe Secret
}
\details{
This function returns the information about the corresponding secret, but does not return the secret value
}
