% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_random_password.R
\name{get_random_password}
\alias{get_random_password}
\title{Get Random Password}
\usage{
get_random_password(exclude_characters, exclude_lowercase, exclude_numbers,
  exclude_punctuation, exclude_uppercase, include_space, password_length,
  require_each_included_type, verbose = getOption("verbose", FALSE))
}
\arguments{
\item{string}{(Optional) A string that includes characters that should not be included in the generated password. The default is that all characters from the included sets can be used.}

\item{boolean}{(Optional) Specifies that the generated password should not include lowercase letters. The default if you do not include this switch parameter is that lowercase letters can be included.}

\item{boolean}{(Optional) Specifies that the generated password should not include digits. The default if you do not include this switch parameter is that digits can be included.}

\item{boolean}{(Optional) Specifies that the generated password should not include punctuation characters. The default if you do not include this switch parameter is that punctuation characters can be included.}

\item{boolean}{(Optional) Specifies that the generated password should not include uppercase letters. The default if you do not include this switch parameter is that uppercase letters can be included.}

\item{boolean}{(Optional) Specifies that the generated password can include the space character. The default if you do not include this switch parameter is that the space character is not included.}

\item{int}{(Optional) The desired length of the generated password. The default value if you do not include this parameter is 32 characters.}

\item{boolean}{(Optional) A boolean value that specifies whether the generated password must include at least one of every allowed character type. The default value is True and the operation requires at least one of every character type.}

\item{boolean}{(Optional) TRUE to print out debug information. Defaults to FALSE}
}
\value{
A string with the generated password.
}
\description{
AWS Secrets Manager Get Random Password
}
\details{
Generates a random password of the specified complexity. This operation is intended for use in the Lambda rotation function. Per best practice, we recommend that you specify the maximum length and include every character type that the system you are generating a password for can support.
}
