% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_resource_policy.R
\name{get_resource_policy}
\alias{get_resource_policy}
\title{Get Resource Policy}
\usage{
get_resource_policy(secret_id, verbose = getOption("verbose", FALSE))
}
\arguments{
\item{string}{The ID of the secret whose policy you which to retrieve}

\item{boolean}{TRUE to print out debug information. Defaults to FALSE}
}
\value{
The JSON text of the resource-based policy document
}
\description{
AWS Secrets Manager Get Resource Policy
}
\details{
Retrieves the JSON text of the resource-based policy document that's attached to the specified secret. The JSON request string input and response output are shown formatted with white space and line breaks for better readability.
}
