% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_secret_binary_value.R
\name{get_secret_binary_value}
\alias{get_secret_binary_value}
\title{Get Secret Binary Value}
\usage{
get_secret_binary_value(secret_id, version_id, version_stage,
  verbose = getOption("verbose", FALSE))
}
\arguments{
\item{string}{The ID of the secret to retrieve}

\item{string}{The version ID of the target secret (optional)}

\item{string}{The version stage of the target secret (optional)}

\item{boolean}{TRUE to print out debug information. Defaults to FALSE}
}
\value{
The secret binary represented as a Base64-encoded string.
}
\description{
AWS Secrets Manager Get Secret Binary Value
}
\details{
This function returns the corresponding secret binary represented as a Base64-encoded string.
}
