% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_secret_string_value.R
\name{get_secret_string_value}
\alias{get_secret_string_value}
\title{Get Secret String Value}
\usage{
get_secret_string_value(secret_id, version_id, version_stage,
  verbose = getOption("verbose", FALSE))
}
\arguments{
\item{string}{The ID of the secret to retrieve}

\item{string}{The version ID of the target secret (optional)}

\item{string}{The version stage of the target secret (optional)}

\item{boolean}{TRUE to print out debug information. Defaults to FALSE}
}
\value{
The secret string stored in Secrets Manager
}
\description{
AWS Secrets Manager Get Secret String Value
}
\details{
This function returns the corresponding secret string
}
