% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_secret_value.R
\name{get_secret_value}
\alias{get_secret_value}
\title{Get Secret String Value}
\usage{
get_secret_value(secret_id, version_id, version_stage,
  verbose = getOption("verbose", FALSE))
}
\arguments{
\item{string}{The ID of the secret to retrieve}

\item{string}{The version ID of the target secret (optional)}

\item{string}{The version stage of the target secret (optional)}

\item{boolean}{TRUE to print out debug information. Defaults to FALSE}
}
\value{
The secret and related fields returned by the Secrets Manager API
}
\description{
AWS Secrets Manager Get Secret Value
}
\details{
This function returns the corresponding secret string
}
