% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_secret_version_ids.R
\name{list_secret_version_ids}
\alias{list_secret_version_ids}
\title{List Secret Version Ids}
\usage{
list_secret_version_ids(secret_id, include_deprecated, max_results,
  next_token, verbose = getOption("verbose", FALSE))
}
\arguments{
\item{string}{The identifier for the secret containing the versions you want to list. You can specify either the Amazon Resource Name (ARN) or the friendly name of the secret.}

\item{boolean}{(Optional) Specifies that you want the results to include versions that do not have any staging labels attached to them. Such versions are considered deprecated and are subject to deletion by Secrets Manager as needed.}

\item{int}{(Optional) Limits the number of results that you want to include in the response.}

\item{string}{(Optional) Used for retrieving results beyond the first page. Use the NextToken value from the previous response to get the next page of data.}

\item{boolean}{TRUE to print out debug information. Defaults to FALSE}
}
\value{
A lists of all of the versions attached to the specified secret.
}
\description{
AWS Secrets Manager List Secret Version Ids
}
\details{
Lists all of the versions attached to the specified secret. The output does not include the SecretString or SecretBinary fields. By default, the list includes only versions that have at least one staging label in VersionStage attached.
}
