% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_secrets.R
\name{list_secrets}
\alias{list_secrets}
\title{List Secrets}
\usage{
list_secrets(max_results, next_token, verbose = getOption("verbose",
  FALSE))
}
\arguments{
\item{int}{(Optional) Limits the number of results that you want to include in the response.}

\item{string}{(Optional) Used for retrieving results beyond the first page. Use the NextToken value from the previous response to get the next page of data.}

\item{boolean}{TRUE to print out debug information. Defaults to FALSE}
}
\value{
A list of secrets without the actual secrets themselves
}
\description{
AWS Secrets Manager List Secrets
}
\details{
Lists all of the secrets that are stored by Secrets Manager in the AWS account. The encrypted fields SecretString and SecretBinary are not included in the output. If additional items exist beyond the maximum you specify, the NextToken response element is present and has a value (isn't null). Include that value as the NextToken request parameter in the next call to the operation to get the next part of the results. Note that Secrets Manager might return fewer results than the maximum even when there are more results available. You should check NextToken after every operation to ensure that you receive all of the results.
}
