% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/put_resource_policy.R
\name{put_resource_policy}
\alias{put_resource_policy}
\title{Put Resource Policy}
\usage{
put_resource_policy(resource_policy, secret_id,
  verbose = getOption("verbose", FALSE))
}
\arguments{
\item{string}{A JSON-formatted string that's constructed according to the grammar and syntax for an AWS resource-based policy. The policy in the string identifies who can access or manage this secret and its versions.}

\item{string}{Specifies the secret that you want to attach the resource-based policy to. You can specify either the ARN or the friendly name of the secret.}

\item{boolean}{TRUE to print out debug information. Defaults to FALSE}
}
\value{
The ARN and name of the new policy
}
\description{
AWS Secrets Manager Put Resource Policy
}
\details{
Attaches the contents of the specified resource-based permission policy to a secret. A resource-based policy is optional.
}
