% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/put_secret_string_value.R
\name{put_secret_string_value}
\alias{put_secret_string_value}
\title{Put Secret String Value}
\usage{
put_secret_string_value(secret_id, secret_string, version_stages,
  verbose = getOption("verbose", FALSE))
}
\arguments{
\item{string}{Specifies the secret to which you want to add a new version. You can specify either the Amazon Resource Name (ARN) or the friendly name of the secret. The secret must already exist.}

\item{string}{The secret text to be stored and encrypted. To encrypt multiple values, it is recommended to use use a JSON text string argument and specify key/value pairs.}

\item{array}{(Optional) An array of strings. Specifies a list of staging labels that are attached to this version of the secret. These staging labels are used to track the versions through the rotation process by the Lambda rotation function.}

\item{boolean}{TRUE to print out debug information. Defaults to FALSE}
}
\value{
The meta data about the secret stored in Secrets Manager
}
\description{
AWS Secrets Manager Put Secret String Value
}
\details{
Stores a new encrypted secret value in the specified secret. To do this, the operation creates a new version and attaches it to the secret.
}
