% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restore_secret.R
\name{restore_secret}
\alias{restore_secret}
\title{Restore Secret}
\usage{
restore_secret(secret_id, verbose = getOption("verbose", FALSE))
}
\arguments{
\item{string}{The ID of the secret to restore}

\item{boolean}{TRUE to print out debug information. Defaults to FALSE}
}
\value{
The meta data about the secret stored in Secrets Manager
}
\description{
AWS Secrets Manager Restore Secret
}
\details{
Cancels the scheduled deletion of a secret by removing the DeletedDate time stamp. This makes the secret accessible to query once again.
}
