% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_resource.R
\name{tag_resource}
\alias{tag_resource}
\title{Tag Resource}
\usage{
tag_resource(secret_id, tags, verbose = getOption("verbose", FALSE))
}
\arguments{
\item{string}{The ID of the secret to retrieve}

\item{dataframe}{A dataframe with columns Key and Value}

\item{boolean}{TRUE to print out debug information. Defaults to FALSE}
}
\value{
The meta data about the secret stored in Secrets Manager (Name, ARN, Version)
}
\description{
AWS Tag Resource
}
\details{
Attaches one or more tags, each consisting of a key name and a value, to the specified secret. Tags are part of the secret's overall metadata, and are not associated with any specific version of the secret. This operation only appends tags to the existing list of tags.
}
